backend/requirements.txt
# Talkify Backend Dependencies

# Core Flask
flask>=2.3.0
flask-sqlalchemy>=3.1.0
flask-jwt-extended>=4.6.0
flask-cors>=4.0.0
flask-migrate>=4.0.0

# Firebase (for authentication sync)
firebase-admin>=6.0.0

# Environment and Config
python-dotenv>=1.0.0

# Validation and Serialization
marshmallow>=3.20.0

# Date/Time utilities
python-dateutil>=2.8.0

# Security
werkzeug>=2.3.0

# Development
pytest>=7.4.0
requests>=2.31.0
